Change `group` to `item` and allow editing on path column
The main reason for this is that for list based values like
```
...
    "var": [0, 1, 2]
...
```
the users would need to find each values manually with the scanner
but by enabling editing/add the path manually, the user can just
insert an item and add the path by themselves.

Context: originally the path is not editable so that the user
wouldn't mistakenly edit it.